add common scripts from eks-distro

+++ build/lib/common.sh
@@ -88,7 +88,7 @@ function build::common::upload_artifacts() {
 function build::gather_licenses() {
   if ! command -v go-licenses &> /dev/null
   then
-    echo " go-licenses not found.  If you need license or attribtuion file handling"
+    echo " go-licenses not found.  If you need license or attribution file handling"
     echo " please refer to the doc in docs/development/attribution-files.md"
     exit
   fi
@@ -96,30 +96,29 @@ function build::gather_licenses() {
   local -r outputdir=$1
   local -r patterns=$2
 
-  # reset the gopath change to make sure and always use
-  # the latest go for generating deps list
-  # older versions behave differently in some cases
-  build::common::remove_go_path
+  # use go 1.16 since 1.17 seems to be more agressive about wanting to update
+  # the go.mod/sum file
+  build::common::use_go_version 1.16
 
   # Force deps to only be pulled form vendor directories
   # this is important in a couple cases where license files
   # have to be manually created
   export GOFLAGS=-mod=vendor
   # force platform to be linux to ensure all deps are picked up
-  export GOOS=linux
-  export GOARCH=amd64
+  export GOOS=linux 
+  export GOARCH=amd64 
 
   mkdir -p "${outputdir}/attribution"
-  # attribution file generated uses the output go-deps and go-license to gather the neccessary
+  # attribution file generated uses the output go-deps and go-license to gather the necessary
   # data about each dependency to generate the amazon approved attribution.txt files
   # go-deps is needed for module versions
   # go-licenses are all the dependencies found from the module(s) that were passed in via patterns
   go list -deps=true -json ./... | jq -s ''  > "${outputdir}/attribution/go-deps.json"
-
+  
   go-licenses save --force $patterns --save_path="${outputdir}/LICENSES"
-
-  # go-licenses can be a bit noisy with its output and lot of it can be confusing
-  # the following messags are safe to ignore since we do not need the license url for our process
+  
+  # go-licenses can be a bit noisy with its output and lot of it can be confusing 
+  # the following messages are safe to ignore since we do not need the license url for our process
   NOISY_MESSAGES="cannot determine URL for|Error discovering URL|unsupported package host"
   go-licenses csv $patterns > "${outputdir}/attribution/go-license.csv" 2>  >(grep -vE "$NOISY_MESSAGES" >&2)
 
@@ -135,21 +134,23 @@ function build::gather_licenses() {
   if cat "${outputdir}/attribution/go-license.csv" | grep -e ",LGPL-" -e ",GPL-"; then
     echo " one of the dependencies is licensed as LGPL or GPL"
     echo " which is prohibited at Amazon"
-    echo " please look into removeing the dependency"
+    echo " please look into removing the dependency"
+    exit 1
   fi
-
 
   # go-license is pretty eager to copy src for certain license types
   # when it does, it applies strange permissions to the copied files
   # which makes deleting them later awkward
   # this behavior may change in the future with the following PR
   # https://github.com/google/go-licenses/pull/28
-  chmod -R 777 "${outputdir}/LICENSES"
+  chmod -R 777 "${outputdir}/LICENSES"  
 
   # most of the packages show up the go-license.csv file as the module name
   # from the go.mod file, storing that away since the source dirs usually get deleted
   MODULE_NAME=$(go mod edit -json | jq -r '.Module.Path')
-  echo $MODULE_NAME > ${outputdir}/attribution/root-module.txt
+  if [ ! -f ${outputdir}/attribution/root-module.txt ]; then
+  	echo $MODULE_NAME > ${outputdir}/attribution/root-module.txt
+  fi
 }
 
 function build::non-golang::gather_licenses(){
@@ -170,32 +171,23 @@ function build::non-golang::gather_licenses(){
 
 function build::generate_attribution(){
   local -r project_root=$1
-  local -r golang_verson=$2
+  local -r golang_version=$2
   local -r output_directory=${3:-"${project_root}/_output"}
   local -r attribution_file=${4:-"${project_root}/ATTRIBUTION.txt"}
 
   local -r root_module_name=$(cat ${output_directory}/attribution/root-module.txt)
-  local -r go_path=$(build::common::get_go_path $golang_verson)
+  local -r go_path=$(build::common::get_go_path $golang_version)
   local -r golang_version_tag=$($go_path/go version | grep -o "go[0-9].* ")
 
   if cat "${output_directory}/attribution/go-license.csv" | grep -e ",LGPL-" -e ",GPL-"; then
     echo " one of the dependencies is licensed as LGPL or GPL"
     echo " which is prohibited at Amazon"
-    echo " please look into removeing the dependency"
+    echo " please look into removing the dependency"
     exit 1
   fi
 
-  generate-attribution $root_module_name $project_root $golang_version_tag $output_directory
+  generate-attribution $root_module_name $project_root $golang_version_tag $output_directory 
   cp -f "${output_directory}/attribution/ATTRIBUTION.txt" $attribution_file
-}
-
-function build::common::remove_go_path() {
-  # This is the path where the specific go binary versions reside in our builder-base image
-  export PATH=$(echo "$PATH" | sed -e "s/^\/go\/go[0-9\.]*\/bin://")
-
-  # This is the path that will most likely be correct if running locally
-  local -r quoted_go_path=$(build::common::re_quote $GOPATH)
-  export PATH=$(echo "$PATH" | sed -e "s/^$quoted_go_path\/go[0-9\.]*\/bin://")
 }
 
 function build::common::get_go_path() {
@@ -242,6 +234,17 @@ function build::common::use_go_version() {
   echo "Adding $gobinarypath to PATH"
   # Adding to the beginning of PATH to allow for builds on specific version if it exists
   export PATH=${gobinarypath}:$PATH
+  export GOCACHE=$(go env GOCACHE)/$version
+}
+
+# Use a seperate build cache for each project/version to ensure there are no
+# shared bits which can mess up the final checksum calculation
+# this is mostly needed for create checksums locally since in the builds
+# different versions of the same project are not built in the same container
+function build::common::set_go_cache() {
+  local -r project=$1
+  local -r git_tag=$2
+  export GOCACHE=$(go env GOCACHE)/$project/$git_tag
 }
 
 function build::common::re_quote() {
